debug: Trying to work out why rendering fails
Renders ok with quarto-1.3 but not with quarto-1.{5,6}-*

Its a problem with a table somewhere, if we remove all code chunks the text and abstract render fine.

Verbose output with...

```
QUARTO_LOG_LEVEL=DEBUG ~/.local/bin/quarto preview
```

And its failing around the code block with label `tbl-ultrasound-investigations-gtsummary`, but can't see how or why.